// services/inventory-service/src/common/constants/role.enum.ts
export enum Role {
  BUYER = 'BUYER',
  CUSTOMER = 'CUSTOMER',
  SELLER = 'SELLER',
  ADMIN = 'ADMIN',
  MODERATOR = 'MODERATOR',
  SUPPORT = 'SUPPORT',
  WAREHOUSE = 'WAREHOUSE',
  SUPER_ADMIN = 'SUPER_ADMIN',
  SERVICE = 'SERVICE'
}

export const STAFF_ROLES: Role[] = [Role.ADMIN, Role.SUPPORT, Role.WAREHOUSE, Role.SELLER, Role.SUPER_ADMIN];